New Singleton implemented, bullets are now counted properly.
