+++ RWM-P2-TEAM-C/Assets/Scripts/Server/AnalyticsManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 06222818e23649c4698ad03d84f415b7
+guid: 9be84d1ae0672514bb533ae1645cad89
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
